prevent users from choosing same course in both modals
