+++ src/pages/TimeTableScheduler.tsx
@@ -16,7 +16,7 @@ import { CheckedCourse, Course, CourseOption, CourseSchedule, Major, MultipleOpt
 import { useShowGrid, useMajor, useCourses } from '../hooks'
 import SelectionExtraCoursesModal from '../components/planner/SelectionExtraCoursesModal'
 
-export const removeDuplicatesFromCourseOption = (courses: CourseOption[]) => {
+export const removeDuplicatesFromCourseOption = (courses: CourseOption[]): CourseOption[] => {
   let frequency: Map<number, number> = new Map()
   let newCourseOptions: CourseOption[] = []
 
@@ -30,8 +30,45 @@ export const removeDuplicatesFromCourseOption = (courses: CourseOption[]) => {
   return newCourseOptions
 }
 
-export const is_null_or_undefined = (course) => {
-  return course === undefined || course === null
+export const removeDuplicatesFromCourseArray = (courses: CheckedCourse[]): CheckedCourse[] => {
+  let frequency: Map<number, number> = new Map()
+  let newCourses: CheckedCourse[] = []
+
+  for (let course of courses) {
+    if (!frequency.has(course.info.id)) {
+      newCourses.push(course)
+      frequency.set(course.info.id, 1)
+    }
+  }
+
+  return newCourses
+}
+
+export const is_null_or_undefined = (element) => {
+  return element === undefined || element === null
+}
+
+/**
+   * This method serves to go to a group of checkboxes and put the correct checked value on the group checkbox
+   */
+ export const controlCoursesGroupCheckbox = (courses: CheckedCourse[], groupCheckboxId: string) => {
+  let some = courses.some((course) => course.checked)
+  let every = courses.every((course) => course.checked)
+
+  //@ts-ignore
+  let checkbox: HTMLInputElement = document.getElementById(groupCheckboxId)
+  if (!checkbox) return
+
+  if (every) {
+    checkbox.checked = true
+    checkbox.indeterminate = false
+  } else if (some) {
+    checkbox.checked = false
+    checkbox.indeterminate = true
+  } else {
+    checkbox.checked = false
+    checkbox.indeterminate = false
+  }
 }
 
 const TimeTableSchedulerPage = () => {
@@ -113,9 +150,11 @@ const TimeTableSchedulerPage = () => {
   const [majors, setMajors] = useState<Major[]>([]) // all the [majors]]]
   const [showGrid, setShowGrid] = useShowGrid() // show the schedule grid or not
   const [checkedCourses, setCheckedCourses] = useCourses("niaefeup-tts.courses") // courses for the major with frontend properties
-  const [extraCoursesActive, setExtraCoursesActive] = useState<boolean>(false)
   const [selectionModalCoursesBuffer, setSelectionModalCoursesBuffer] = useCourses("niaefeup-tts.courses-buffer")
-  const [extraCourseModalCoursesBuffer, setExtraCourseModalCoursesBuffer] = useCourses("niaefeup-tts.extra-courses")
+  const [extraCoursesModalBuffer, setExtraCoursesModalBuffer] = useCourses("niaefeup-tts.extra-courses-buffer")
+  const [extraMajorEqualToMainMajor, setExtraMajorEqualToMainMajor] = useState<boolean>(false)
+  const [chosenMajorMainModalEqualToExtra, setChosenMajorMainModalEqualToExtra] = useState<boolean>(false)
+  const [extraCoursesActive, setExtraCoursesActive] = useState<boolean>(false)
   const [multipleOptions, setMultipleOptions] = useState<MultipleOptions>(
     { index: 0, selected: [], options: [], names: Array.from({ length: 10 }, (_, i) => `Horário ${i + 1}`) }
   ) // schedule options and selected schedule
@@ -144,7 +183,7 @@ const TimeTableSchedulerPage = () => {
 
 
   const getCoursesForMajor = (major: Major, majorChangedRef) => {
-    if (major === null || (majorChangedRef.current === false && checkedCourses.length > 0)) return
+ if (major === null || (majorChangedRef.current === false && checkedCourses.length > 0)) return
 
     let finalNewCheckedCourses: CheckedCourse[][]
 
@@ -157,14 +196,9 @@ const TimeTableSchedulerPage = () => {
     BackendAPI.getCourses(major).then((courses: Course[]) => {
       const majorCourses = groupMajorCoursesByYear(courses)
       const newCheckedCourses = courseToCheckedCourse(majorCourses)
+      let uCC = updateCheckedCourses(newCheckedCourses, multipleOptions.selected)
       majorChangedRef.current = false
-
-
-      for (let courses of newCheckedCourses) {
-        finalNewCheckedCourses.push(courses)
-      }
-
-      setCheckedCourses([...finalNewCheckedCourses])
+      setCheckedCourses([...uCC])
     })
   }
 
@@ -186,8 +220,21 @@ const TimeTableSchedulerPage = () => {
     getCoursesForMajor(major, majorChangedRef)
   }, [major, majorChangedRef, checkedCourses, setCheckedCourses])
 
+  /**
+   * Checks if the current selected extra major is the same as the main major selected in the selectionModal
+   * If it is not, it will fetch the courses for the current major
+   */
   useEffect(() => {
-    getCoursesForMajor(extraCoursesMajor, extraCoursesMajorChangedRef)
+    if(is_null_or_undefined(extraCoursesMajor)) return
+
+    if(extraCoursesMajor.acronym === major.acronym && !chosenMajorMainModalEqualToExtra) {
+      setExtraMajorEqualToMainMajor(true)
+    } else {
+      getCoursesForMajor(extraCoursesMajor, extraCoursesMajorChangedRef)
+
+      if(extraMajorEqualToMainMajor) setExtraMajorEqualToMainMajor(false)
+    }
+    
   }, [extraCoursesMajor, extraCoursesMajorChangedRef, checkedCourses, setCheckedCourses])
 
   const updateCheckedCourses = (newCheckedCourses: CheckedCourse[][], importedCourses: CourseOption[]) => {
@@ -215,30 +262,10 @@ const TimeTableSchedulerPage = () => {
     return newCheckedCourses
   }
 
-  // once a major has been picked => fetch courses for the major
-
-  useEffect(() => {
-    if (major === null || (majorChangedRef.current === false && checkedCourses.length > 0 && !isImportedSchedule)) {
-      return
-    }
-    BackendAPI.getCourses(major).then((courses: Course[]) => {
-      const majorCourses = groupMajorCoursesByYear(courses)
-      const newCheckedCourses = courseToCheckedCourse(majorCourses)
-      let uCC = updateCheckedCourses(newCheckedCourses, multipleOptions.selected)
-      majorChangedRef.current = false
-      if (isImportedSchedule)
-        setCheckedCourses([...uCC])
-      else
-        setCheckedCourses([checkedCourses[0], ...uCC.slice(1)])
-
-    })
-
-    // this line is needed to since adding isImportedSchedule and SetCheckedCourses to the dependencies array would cause Import not to work
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [major, majorChangedRef, checkedCourses, multipleOptions])
-
   // fetch schedules for the courses and preserve course options (once courses have been picked)
   useEffect(() => {
+    // setMultipleOptions({ index: 0, selected: [], options: [] })
+
     let pickedCourses = getPickedCourses(checkedCourses)
     if (pickedCourses.length === 0) return
 
@@ -264,11 +291,14 @@ const TimeTableSchedulerPage = () => {
       return value ? { shown: value.shown, option: value.option } : { shown: { T: true, TP: true }, option: null }
     }
 
+    let isNewerPromise = true
+    
+    // CHANGE
     if(isExtraUcsModelOpen) {
       pickedCourses = pickedCourses.concat(selectionModalCoursesBuffer.flat().filter(course => course.checked))
+      pickedCourses = removeDuplicatesFromCourseArray(pickedCourses)
     }
 
-    let isNewerPromise = true;
     fetchPickedSchedules(pickedCourses).then((schedules: CourseSchedule[][]) => {
       
       if (isNewerPromise) {
@@ -394,19 +424,22 @@ const TimeTableSchedulerPage = () => {
               extraCoursesActiveHook={[extraCoursesActive, setExtraCoursesActive]}
               extraCoursesModalOpenHook={[isExtraUcsModelOpen, setIsExtraUcsModalOpen]}
               sourceBufferHook={[selectionModalCoursesBuffer, setSelectionModalCoursesBuffer]}
-              destBufferHook={[extraCourseModalCoursesBuffer, setExtraCourseModalCoursesBuffer]}
+              destBufferHook={[extraCoursesModalBuffer, setExtraCoursesModalBuffer]}
+              repeatedCourseControlHook={[chosenMajorMainModalEqualToExtra, setChosenMajorMainModalEqualToExtra]}
             />
+
             {isExtraUcsModelOpen ?
               <SelectionExtraCoursesModal
                 majors={majors}
                 openHook={[isExtraUcsModelOpen, setIsExtraUcsModalOpen]}
                 majorHook={[extraCoursesMajor, setExtraCoursesMajor]}
                 coursesHook={[checkedCourses, setCheckedCourses]}
-                sourceBufferHook={[extraCourseModalCoursesBuffer, setExtraCourseModalCoursesBuffer]}
+                sourceBufferHook={[extraCoursesModalBuffer, setExtraCoursesModalBuffer]}
                 destBufferHook={[selectionModalCoursesBuffer, setSelectionModalCoursesBuffer]}
                 multipleOptionsHook={[multipleOptions, setMultipleOptions]}
+                repeatedCourseControlHook={[extraMajorEqualToMainMajor, setExtraMajorEqualToMainMajor]}
               />
-              : <></>}
+              : <></> }
             <MoreActionsButton
               schedule={multipleOptions.selected}
               showGridHook={[showGrid, setShowGrid]}
